Make editing/adding badges work with Vanilla 2.3
Feature was broken because the property $InputPrefix is not supported in current Vanilla versions.
As a result the form fields that belong to the rule weren't recognized as criteria fields and reported as "missing".
I solved that problem by listing the general form fields and taking the difference between those and the form fields from the postback as the criteria fields.